bsp: compact tree node memory representation
Previously, the memory sizes of split and leaf nodes were
significantly different, yet they were confined to a single variant.
This setup led to substantial memory inefficiency, with a structure
comprising six number types and one data type.

To address this issue, the leaf node's bound and data components have
now been moved to separate slab storage. This change allows for a more
compact representation of heavy data in memory. Additionally, it
narrows the memory gap between split and leaf nodes, resulting in a
more compact `nodes` container.

This improvement was essential as the tree expanded. Typically, the
ratio between split and leaf nodes approached 1:1, exacerbating the
inefficiency of the memory gap.